Move TablePermissions to ApiKey
